LPS-74724 Move comparator to util.comparator package
